backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
PyMySQL>=1.1.0
cryptography>=41.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.0.0
httplib2>=0.20.0
requests>=2.31.0
python-multipart>=0.0.6
email_validator>=2.0.0